requirements.txt
python-dotenv
fastapi
requests
pydantic
uvicorn